华为云

+++ cloud-service/cloud-system/src/main/java/com/cloud/system/controller/SysOssController.java
@@ -5,16 +5,14 @@ import com.cloud.common.auth.annotation.HasPermissions;
 import com.cloud.common.core.controller.BaseController;
 import com.cloud.common.core.domain.R;
 import com.cloud.common.exception.file.OssException;
+import com.cloud.common.utils.StringUtils;
 import com.cloud.common.utils.ValidatorUtils;
+import com.cloud.common.utils.file.FileUtils;
 import com.cloud.system.domain.entity.SysOss;
-import com.cloud.system.oss.CloudConstant;
+import com.cloud.system.oss.*;
 import com.cloud.system.oss.CloudConstant.CloudService;
-import com.cloud.system.oss.CloudStorageConfig;
-import com.cloud.system.oss.CloudStorageService;
-import com.cloud.system.oss.OSSFactory;
 import com.cloud.system.oss.valdator.AliyunGroup;
-import com.cloud.system.oss.valdator.QcloudGroup;
-import com.cloud.system.oss.valdator.QiniuGroup;
+import com.cloud.system.oss.valdator.HuaweiGroup;
 import com.cloud.system.service.ISysConfigService;
 import com.cloud.system.service.ISysOssService;
 import com.google.gson.Gson;
@@ -62,15 +60,12 @@ public class SysOssController extends BaseController {
     public R saveConfig(CloudStorageConfig config) {
         // 校验类型
         ValidatorUtils.validateEntity(config);
-        if (config.getType() == CloudService.QINIU.getValue()) {
-            // 校验七牛数据
-            ValidatorUtils.validateEntity(config, QiniuGroup.class);
+        if (config.getType() == CloudService.HUAWEIYUN.getValue()) {
+            // 校验华为云数据
+            ValidatorUtils.validateEntity(config, HuaweiGroup.class);
         } else if (config.getType() == CloudService.ALIYUN.getValue()) {
             // 校验阿里云数据
             ValidatorUtils.validateEntity(config, AliyunGroup.class);
-        } else if (config.getType() == CloudService.QCLOUD.getValue()) {
-            // 校验腾讯云数据
-            ValidatorUtils.validateEntity(config, QcloudGroup.class);
         }
         return toAjax(sysConfigService.updateValueByKey(KEY, new Gson().toJson(config)));
     }
@@ -129,11 +124,43 @@ public class SysOssController extends BaseController {
     }
 
     /**
+     * 下载文件
+     *
+     */
+    @PostMapping("downLoad")
+    public void downLoad(String url,String fileName) throws IOException {
+        String realName = new String();
+        if(StringUtils.isNotBlank(fileName)){
+            realName=fileName;
+        }else{
+            realName = url;
+        }
+        CloudStorageService storage = OSSFactory.build();
+
+        getResponse().setCharacterEncoding("utf-8");
+        // 下载使用"application/octet-stream"更标准
+        getResponse().setContentType("application/octet-stream");
+        getResponse().setHeader("Content-Disposition",
+                "attachment;filename=" + FileUtils.setFileDownloadHeader(getRequest(), realName));
+        storage.downLoad(url,getResponse().getOutputStream());
+    }
+
+    /**
      * 删除文件上传
      */
     @PostMapping("remove")
     @HasPermissions("sys:oss:remove")
     public R remove(String ids) {
+        if(ids==null){
+            return R.error("参数为空!!");
+        }
+        CloudStorageService storage = OSSFactory.build();
+        for(String id:ids.split(",")){
+            SysOss sysOss = sysOssService.selectSysOssById(Long.parseLong(id));
+            storage.deleteFile(sysOss.getUrl());
+        }
         return toAjax(sysOssService.deleteSysOssByIds(ids));
     }
+
+
 }
